upipe-hbrmt: change pixel argument types to uintptr_t
The reasoning for this is:
* We want it to be unsigned because we don't expect a negative value to
  be passed.  We perform unsigned shifts on the value in some cases.
* We want it to be register sized so that we do not have to extend it on
  x86-64.

+++ lib/upipe-hbrmt/upipe_pack10bit.c
@@ -78,7 +78,7 @@ struct upipe_pack10bit {
     struct uchain blockers;
 
     /** packing */
-    void (*pack)(uint8_t *dst, const uint8_t *y, int64_t pixels);
+    void (*pack)(uint8_t *dst, const uint8_t *y, uintptr_t pixels);
 
     /** public upipe structure */
     struct upipe upipe;
